+++ src/lib/signing.rs
@@ -13,16 +13,6 @@ use std::time::Duration;
 use super::ledger::LedgerIdentity;
 use super::{get_agent, get_default_role};
 
-#[derive(Debug)]
-pub struct MessageError(String);
-
-impl std::fmt::Display for MessageError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", &self.0)
-    }
-}
-impl std::error::Error for MessageError {}
-
 /// Represents a signed message with the corresponding request id.
 #[derive(Clone)]
 pub struct SignedMessageWithRequestId {